line-notify-web/requirements.txt
Flask==1.1.2
requests==2.24.0
python-dotenv==0.19.0
urllib3==1.25.9
